streamlit/pyproject.toml
[project]
name = "streamlit_app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.28.0",
]
